+++ CHANGELOG.md
@@ -11,6 +11,12 @@ The REST API versions in the table are just for your information as the values a
 |\>= 10.0.0|3
 |<= 9.0.1|2
 
+<a name="11.12.0"></a>
+## 2019-06-04 Version [11.12.0](https://github.com/gooddata/gooddata-js/compare/v11.11.0...v11.12.0)
+
+- show error when exporting the Restricted insight
+- add showFilters to IExportConfig, this will add the applied filters to the exported file.
+
 <a name="11.11.0"></a>
 ## 2019-05-21 Version [11.11.0](https://github.com/gooddata/gooddata-js/compare/v11.10.0...v11.11.0)